finds the path ans changed simulation
doesn't follow the path yet. need to implement that - probably super hard cause of turns. A* search algorithm(sort of: more like heauristic search) for pathfinding , important to know that it's not always the fastest route - easy fix but takes a little more coding.(can do it last)
added line to simulation to fix memory leak